Parametrize quiche lib tests by congestion control algorithm (#2013)
* parametrize quiche lib tests by congestion control algorithm

* Enable dgram_send_app_limited test for bbr2_gcongestion

* Fix multiplication overflow in 32 bit builds

* fix openssl tests

+++ quiche/src/recovery/gcongestion/bbr2.rs
@@ -561,10 +561,14 @@ impl CongestionControl for BBRv2 {
     }
 
     fn update_mss(&mut self, new_mss: usize) {
-        self.cwnd_limits.hi = self.cwnd_limits.hi * new_mss / self.mss;
-        self.cwnd_limits.lo = self.cwnd_limits.lo * new_mss / self.mss;
-        self.cwnd = self.cwnd * new_mss / self.mss;
-        self.initial_cwnd = self.initial_cwnd * new_mss / self.mss;
+        self.cwnd_limits.hi = (self.cwnd_limits.hi as u64 * new_mss as u64 /
+            self.mss as u64) as usize;
+        self.cwnd_limits.lo = (self.cwnd_limits.lo as u64 * new_mss as u64 /
+            self.mss as u64) as usize;
+        self.cwnd =
+            (self.cwnd as u64 * new_mss as u64 / self.mss as u64) as usize;
+        self.initial_cwnd = (self.initial_cwnd as u64 * new_mss as u64 /
+            self.mss as u64) as usize;
         self.mss = new_mss;
     }
